crates/sel4/config: Merge generic into macros and config
