+++ crates/sel4/config/Cargo.nix
@@ -16,11 +16,13 @@ mk {
   build-dependencies = {
     inherit (versions)
       prettyplease
+      proc-macro2
+      quote
     ;
     syn = { version = versions.syn; features = [ "parsing" ]; };
     inherit (localCrates)
       sel4-config-data
-      sel4-config-generic
+      sel4-config-generic-types
     ;
   };
 }
