--- crates/sel4/config/generic/Cargo.toml
@@ -1,24 +1,0 @@
-#
-# Copyright 2023, Colias Group, LLC
-#
-# SPDX-License-Identifier: BSD-2-Clause
-#
-#
-# This file is generated from './Cargo.nix'. You can edit this file directly
-# if you are not using this project's Cargo manifest management tools.
-# See 'hacking/cargo-manifest-management/README.md' for more information.
-#
-
-[package]
-name = "sel4-config-generic"
-version = "0.1.0"
-authors = ["Nick Spinale <nick.spinale@coliasgroup.com>"]
-edition = "2021"
-license = "BSD-2-Clause"
-
-[dependencies]
-fallible-iterator = "0.3.0"
-proc-macro2 = "1.0.89"
-quote = "1.0.37"
-sel4-config-generic-types = { path = "types" }
-syn = { version = "2.0.85", features = ["full"] }
