--- crates/sel4/config/generic/src/generate_consts.rs
@@ -1,35 +1,0 @@
-//
-// Copyright 2024, Colias Group, LLC
-//
-// SPDX-License-Identifier: BSD-2-Clause
-//
-
-use proc_macro2::TokenStream;
-use quote::{format_ident, quote};
-
-use sel4_config_generic_types::{Configuration, Value};
-
-pub fn generate_consts(config: &Configuration) -> TokenStream {
-    let items = config.iter().map(|(k, v)| {
-        let k = format_ident!("{}", k);
-        let tv = match v {
-            Value::Bool(v) => {
-                quote! {
-                    bool = #v
-                }
-            }
-            Value::String(v) => {
-                quote! {
-                    &str = #v
-                }
-            }
-        };
-        quote! {
-            pub const #k: #tv;
-        }
-    });
-
-    quote! {
-        #(#items)*
-    }
-}
