+++ crates/sel4/config/macros/Cargo.toml
@@ -20,5 +20,9 @@ license = "BSD-2-Clause"
 proc-macro = true
 
 [dependencies]
+fallible-iterator = "0.3.0"
+proc-macro2 = "1.0.89"
+quote = "1.0.37"
 sel4-config-data = { path = "../data" }
-sel4-config-generic = { path = "../generic" }
+sel4-config-generic-types = { path = "../generic/types" }
+syn = { version = "2.0.85", features = ["full"] }
