+++ crates/sel4/config/macros/src/generic/common_helpers.rs
