+++ crates/sel4/config/macros/src/lib.rs
@@ -9,7 +9,10 @@ use std::marker::PhantomData;
 use proc_macro::TokenStream;
 
 use sel4_config_data::get_kernel_config;
-use sel4_config_generic::MacroImpls;
+
+mod generic;
+
+use generic::MacroImpls;
 
 fn get_impls() -> MacroImpls<'static> {
     MacroImpls::new(get_kernel_config(), "sel4_cfg")
